requirements.txt
pandas>=2.1
openpyxl>=3.1
requests>=2.31
beautifulsoup4>=4.12
python-docx>=1.1
PyMuPDF>=1.24
Pillow>=10.0
pywin32>=306; platform_system=="Windows"
